- 재고 조회 및 감소 로직에서 발생하는 동시성 제어를 위한 실습코드 추가(3) - RabbitMQ 도입(rdb와의 동기화 작업을 수행하기 위해 rabbitMQ를 통해 비동기 로직 추가)   - RabbitMq 의존성 및 yml 세팅 추가   - RabbitMqConfig 세팅 (연결객체와 생성템플릿객체 생성)   - RabbitMqStockService 세팅(publish: 발행할 채널에 대해 메시지 발행, subscribe : 큐를 바라보며 발행된 메시지가 있으면 호출됨)

+++ src/main/java/com/beyond/order_system/ordering/dto/request/OrderCreateReqDto.java
@@ -26,6 +26,6 @@ public class OrderCreateReqDto {
 
         @NotNull
         @Min(1)
-        private Integer productCount;
+        private Long productCount;
     }
 }
